mark integer-compared-to-float test as TODO on 32 bit machines
where ivsize < 8, the integer is being represented as a PV rather than an IV;
where nvsize < 8, the float is being represented as a PV rather than an NV;
if one of these comes through as a string, comparison with the other will
fail on the types being different; if both come through as a string, the
string comparison will fail.

e.g. see GHI #10 and
http://www.cpantesters.org/cpan/report/d9ff55c4-a13d-11ea-9066-e374b0ba08e8
http://www.cpantesters.org/cpan/report/a1898f1c-a097-11ea-b276-2fafcccccfd6
http://www.cpantesters.org/cpan/report/f6457db7-6bf7-1014-9edb-f53f8c633b70