allow more usecases for $recursiveRef resolution
$recursiveRef having a value anything other than "#" is not well-defined
anyway (and this is being fixed in draft 2020-xx), but we only run into
trouble when both the base uri (from the outermost $recursiveAnchor) and the
final target URI (from $recursiveRef) both have non-empty fragments.

This allows the use of $recursiveAnchor from anywhere an "$id" is not
explicitly provided, such as inside a $defs property.