clean up what we call an "exception"
Anything that can be found by statically inspecting the schema(s) is not an
exception, but just an error, even if we bail out immediately upon finding it.
Exceptions are only things we find at runtime, that can be resolved by
changing something else (e.g. making a resource available at an indicated URI,
or changing a configuration value), or a bona fide programming error (!)