delay looking up vocabulary classes until actually using the metaschema
e.g. when using a metaschema as a subschema via $ref, in which case its
$vocabulary entries are not used, so the vocabulary URIs do not need to be
implemented as classes.

This also means we no longer store the metaschema vocabulary classes on the
document object itself -- we delay calculating these until traversing a schema
that references this schema as a metaschema, where we can cache this information
on the evaluator itself.

Also fix the use of a metaschema that doesn't use the $vocabulary keyword at all
(e.g. when making a certain keyword definition more strict, or adding syntax
to annotation-only keywords).