performance enhancement: store annotations in hashref form in $state
This lets us skip calling the ::Annotation constructor, as well as
performing some operations with Mojo::URLs, until we know they are needed.
Much of the time this data is thrown away.

We don't do the same thing for errors, because traverse() is a public method
that is also subclassed elsewhere, that returns errors as part of its $state
return value; also errors, when generated, are more often used rather than
discarded.