do not collect annotations when not needed by an unevaluated* keyword
..or explicitly requested, of course.

At traverse time we used to take note of whether we saw an unevaluated*
keyword _anywhere_ in the schema and collect based on that, but that's wildly
inefficient when evaluating with a large schema that only uses unevaluated* in
a few places (such as the OpenAPI v3.1 specification schema). Now we take
advantage of the knowledge that an annotation coming from a property- or
item-based applicator keyword is only useful at the current data instance
location, and once we've moved on any annotations will not be useful to a
local unevaluated* keyword.