allow dualvars as long as the stringified number matches the string form
This lets us work around an issue in YAML::XS where all numbers are inflated
as SVs with both the PV and NV or IV flags, which we have been interpreting as
"ambiguous type", which therefore causes validation errors when decoding json
schemas from yaml files using YAML::XS.

This also seems to be the behaviour of Cpanel::JSON::XS, where if the two
values match, the value is serialized as a number, and if they differ it is
serialized as a string:

  use Cpanel::JSON::XS;
  use Scalar::Util 'dualvar';
  say encode_json({foo => dualvar(5, '5'});
  say encode_json({foo => dualvar(5, "five")})

  {"foo":5}
  {"foo":"five"}

This gets us closer to 5.36+ behaviour where simply using a number in string
context (where a PV is added to the scalar) does not make the number unusable
in json schemas.  However, it may cause problems with other code downstream
that attempts to use the value and cannot tell if it is a string or number
(e.g. a database driver).