properly handle vocabulary lists changing midway through evaluation
...which can happen when evaluating a local subschema with a $schema keyword
that changes the vocabulary list.  For subschemas evaluated by a $ref, we were
already swapping in the vocabulary list (stored in the resource index)
proactively before iterating over vocabularies and keywords, but when
recursing locally we rely on evaluating the $id keyword, which does not happen
until we have already started iterating through the vocabulary list and are
processing the Core vocabulary.