also handle local subschemas using draft4, where $id is now id
To do this we need to consider $schema ahead of $id, because changing the
metaschema dialect can also change the keyword list (in a material way) for
the Core vocabulary, because identifiers are provided by 'id' in draft4 rather
than '$id'.  Then we also need to recalculate the keyword list if the
specification version changes.