refactor format evaluation so optional modules are loaded closer to where they are needed
This lets us return a "valid" result when formats are invoked against an
unsupported data type (which is "not a string" for all builtin formats),
rather than throwing an exception if an optional module is not installed.

This also gets the draft2019-09/draft2020-12 format.json acceptance tests
passing again (which some smokers run).