remove local use of $document->evaluator
In a subsequent release we will remove this attribute entirely, necessitating
the passing around of an evaluator object whenever it is needed rather than
fetching it from the document object. The impact will be mostly confined to
the OpenAPI-Modern distribution, which uses the document's evaluator much more
heavily in its local document construction.