remove 'configs' from $state and the resource index
This was added in commit d0d13c72 and originally called 'evaluator_configs',
as an attribute on JSMD, to store a 'collect_annotations' override for schemas
containing unevaluated* keywords, in cases where the user hadn't explicitly
requested collect_annotations => 1, and added to the resource index as
'configs'. Later the attribute was renamed to 'evaluation_configs', and then
subsequently folded into $state as 'configs'.

Then the annotation logic got smarter and we stopped needing a
$state-scoped config for it, in commit de7a0c46.  $state->{configs} was never
used for anything else, so it just sat there dormant, until today.