new "with_defaults" feature, to collect data from "default" keywords for missing data
This cannot be implemented with callbacks because for missing data, keywords
are not actually evaluated. We need to do it from within the implementations
of "properties" and "prefixItems" to look for properties or items that have a
specification, but are missing in the instance data.

This option could be used by an OpenAPI linter to provide a list of implicit
defaults for an OAD using various permutations of various objects (e.g.
parameter objects and encoding objects, which have many different defaults or
allowed/disallowed properties depending on what configuration is used).

This can also be done directly with local tooling:

  $ git clone git@github.com:OAI/OpenAPI-Specification.git
  $ cd OpenAPI-Specification
  $ git checkout v3.2-dev
  $ json-schema-eval --with-defaults --schema src/schemas/validation/schema.yaml --data tests/schema/pass/parameter-object-cookie-form-allowReserved.yaml

  {
    "defaults": {
      "/components/parameters/my_cookie/allowReserved": false,
      "/components/parameters/my_cookie/deprecated": false,
      "/components/parameters/my_cookie/explode": true,
      "/components/parameters/my_cookie/required": false,
      "/components/parameters/my_cookie/style": "form",
      "/jsonSchemaDialect": "https://spec.openapis.org/oas/3.2/dialect/WORK-IN-PROGRESS",
      "/servers": [
        {
          "url": "/"
        }
      ]
    },
    "valid": true
  }