fix strict validation of draft4 schemas
So, funny story. I got the bright idea to bundle the OAS 3.0 schema into
OpenAPI-Modern along with the 3.1 and 3.2 schemas so as to allow one to
easily verify 3.0 OpenAPI documents from the command line, like we can do for
3.1 and 3.2 documents (using the json-schema-eval command line utility).
So I set about updating the constants in OpenAPI::Modern::Utilities to be
aware of this file, and then run `update-schemas` to download the file from
spec.openapis.org -- and I get a validation error: a whole bunch of '$ref'
keywords are unknown!  It turns out that http://json-schema.org/draft-04/schema#
doesn't mention the $ref keyword in its list of known properties at all!
So we need to add a special exemption for this.