keep "short_circuit" enabled in more cases
Instead of always turning it off when an "unevaluated*" keyword is present, we
will just only not short-circuit so eagerly in "anyOf" (and "contains", for
draft2020-12 only) evaluation when we would otherwise have had enough passing
subschemas to short-circuit with a success. This allows us to continue to
collect the annotations we need to allow the "unevaluated*" keyword to
succeed, while skipping unnecessary evaluations in the presence of a
guaranteed failed validation.

This option will also no longer get in the way of "strict" mode functioning
properly, when it is enabled.