identify bad references in schemas
We identify all references in the implementation of the '$ref',
'$recursiveRef' and '$dynamicRef' keywords in the Core vocabulary,
and process them in the Document constructor after all identifiers and
entities have been extracted.

Subclasses of the Document object (namely
JSON::Schema::Modern::Document::OpenAPI) can also identify references to
non-schemas through their own parsing mechanisms, which will be processed in
the same way.